+++ src/main/java/com/duokewat/towardscloud/stockservice/json/Result.java
@@ -1,10 +1,12 @@
 package com.duokewat.towardscloud.stockservice.json;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
 public class Result {
-	private Long regularMarketPrice;
+	private BigDecimal regularMarketPrice;
 }
